+++ src/Common/theming/icons.ts
@@ -28,13 +28,14 @@ export type IconName =
 | 'weather-lightning' | 'weather-rain' | 'weather-snow' | 'weather-sunrise' | 'weather-sunset' | 'weather-sunshine' 
 | 'weather-wind' | 'weather-wind-force' | 'window' | 'world' | 'caret-left' | 'fullscreen' | 'justified-text' | 'eye-open' 
 | 'eye-close' | 'cancel' | 'enterprise' | 'card' | 'purchase' | 'notification' | 'stats' | 'acceptation' | 'check' 
-| 'billing' | 'page' | 'circle-plus' | 'circle-close' | 'circle-info' | 'file' | 'enter' | 'profile2' | 'burger-menu2' 
+| 'billing' | 'page' | 'circle-plus' | 'circle-close' | 'circle-info' | 'circle-check' | 'circle-play' | 'file' | 'enter' | 'profile2' | 'burger-menu2' 
 | 'mobile-up' | 'delivery' | 'parcel' | 'store' | 'group' | 'colored-calendar' | 'colored-transaction' | 'colored-equalizer' 
 | 'colored-card' | 'colored-card-inclined' | 'colored-list' | 'colored-summary' | 'pen' | 'bill-sheet' | 'copy' | 'duplicate' 
 | 'file2' | 'home' | 'up-password' | 'condition' | 'on-off' | 'box' | 'box-birth' | 'box-christmas' | 'box-christmas2' 
 | 'box-cross' | 'box-hat' | 'box-heart' | 'box-infinity' | 'box-mustache' | 'box-rings' | 'box-school' | 'box-smile' 
 | 'box-stars' | 'box-trophy' | 'price' | 'letter' | 'post' | 'clear' | 'info' | 'back-arrow' | 'stacking' | 'zendesk' 
-| 'faq' | 'conecs' | 'moveup' | 'eye-open2' | 'file3' | 'invoice' | 'warning' | 'nfc' | 'nfc-off' | 'profile3' | 'group2' | 'padlock';
+| 'faq' | 'conecs' | 'moveup' | 'eye-open2' | 'file3' | 'invoice' | 'warning' | 'nfc' | 'nfc-off' | 'profile3' | 'group2' | 'padlock'
+| 'file-edit' | 'user-circle' | 'safe';
 
 export const IconNames: IconName[] = ['menu', 'previous', 'list', 'gift', 'attachment', 'add', 'asterisk', 'calendar', 'caret-down',
 'caret-right', 'caret-up', 'close', 'comment', 'confirm', 'cross', 'delete', 'download', 'edit', 'email', 'error-sign', 'export', 
@@ -54,12 +55,12 @@ export const IconNames: IconName[] = ['menu', 'previous', 'list', 'gift', 'attac
 'trash-can', 'tv', 'fax', 'video-camera', 'weather-cloud', 'weather-lightning', 'weather-rain', 'weather-snow', 'weather-sunrise', 
 'weather-sunset', 'weather-sunshine', 'weather-wind', 'weather-wind-force', 'window', 'world', 'caret-left', 'fullscreen', 'justified-text', 
 'eye-open', 'eye-close', 'cancel', 'enterprise', 'card', 'purchase', 'notification', 'stats', 'acceptation', 'check', 'billing', 'page', 
-'circle-plus', 'circle-close', 'circle-info', 'file', 'enter', 'profile2', 'burger-menu2', 'mobile-up', 'delivery', 'parcel', 'store', 'group',
+'circle-plus', 'circle-close', 'circle-info', 'circle-check', 'circle-play', 'file', 'enter', 'profile2', 'burger-menu2', 'mobile-up', 'delivery', 'parcel', 'store', 'group',
 'colored-calendar', 'colored-transaction', 'colored-equalizer', 'colored-card', 'colored-card-inclined', 'colored-list', 'colored-summary', 
 'pen', 'bill-sheet', 'copy', 'duplicate', 'file2', 'home', 'up-password', 'condition', 'on-off', 'box', 'box-birth', 'box-christmas', 'box-christmas2', 
 'box-cross', 'box-hat', 'box-heart', 'box-infinity', 'box-mustache', 'box-rings', 'box-school', 'box-smile', 'box-stars', 'box-trophy', 'price',
 'letter', 'post', 'clear','info','back-arrow','stacking', 'zendesk', 'faq', 'conecs', 'moveup', 'eye-open2', 'file3', 'invoice', 'warning', 'nfc', 'nfc-off',
-'profile3', 'group2', 'padlock'] ;
+'profile3', 'group2', 'padlock', 'file-edit', 'user-circle', 'safe'] ;
 
 let dictionary = {};