+++ src/Components/Inputs/DropFile/types.ts
@@ -59,6 +59,10 @@ export interface UpDropFileProps extends WithThemeProps {
     dropLabel?: string;
     separatorLabel?: string;
     allowExtensionsLabel?: string;
+    allowedExtensionsErrorMessage?: (
+      allowedExtensions?: string[],
+      value?: IFile
+    ) => string;
     displaySelectFile?: boolean;
     tabIndex?: number;
     onChange?: (event: React.ChangeEvent<any>, value: IFile) => void;