+++ src/Components/Inputs/Radio/UpRadio.tsx
@@ -113,7 +113,7 @@ class UpRadio extends BaseControlComponent<UpRadioProps, any> {
     }
 
     showError() {
-        return this.props.showError !== undefined ? this.props.showError === true : this.hasError;
+        return this.props.showError !== undefined ? typeof this.props.showError === "function" ? (this.props.showError as Function)(this.state) : this.props.showError === true : this.hasError;
     }
 
     showSuccess() {