+++ src/Components/Inputs/RichText/UpRichText.tsx
@@ -36,7 +36,7 @@ export default class UpRichText extends BaseControlComponent<UpRichTextProps, st
     
     showError() {
         return this.props.showError !== undefined
-            ? this.props.showError === true
+            ? typeof this.props.showError === "function" ? (this.props.showError as Function)(this.state) : this.props.showError === true
             : this.hasError;
     }