+++ src/Components/Inputs/Text/UpText.tsx
@@ -93,7 +93,7 @@ export default class UpText extends BaseControlComponent<UpTextProps, string> {
 
     showError() {
         return this.props.showError !== undefined
-            ? this.props.showError === true
+            ? typeof this.props.showError === "function" ? (this.props.showError as Function)(this.state) : this.props.showError === true
             : this.hasError;
     }